Asyncronous load of status page to improve speed of initial load.
Splits out the main bulk of status page into a separate file that
is either "included" in the main page, or can be asynchronously
loaded using javascript AJAX techniques.  The async load significantly
speeds up the initial appearance of the status page, though over
all time to populate the whole page is largely unchanged.
Defaults to async load, but this can be changed in preferences tab.