Update to algorithm to add A and AAAA DNS records through DNSMASQ.
First use "localise-queries" to only return the interface IP address
of the network that is making the request.  This works for IPv4 only.
So for IPv6 we only add the interface-name for the first interface
that has an IPv6 address assigned to it.  This means that DNS query
on our host name will return only one IPv4 address and/or only
"one set" of IPv6 addresses (a GUA and/or ULA for the interface)